The Basics of HTML--Puntos 6, 7
-Create a "sign up" form with fields for: first name, last name, email,
birthday, a dropdown to choose your favourite sport, and a text-area to
include a small bio for the user. Add a button at the end to submit the
form, and another one to clear the form. Add relevant validation rules
for all fields (like required fields, valid email).
-Test your HTML files in at least Firefox, Chrome, IE, and Chrome for
Android or iOS Safari.