Topic 3: 5
Create a file called app.js using browserify. Add that script to an
index.html. Check that it works opening it in the browser.

I was able to create the file but for some reason when i try to
inplement the API, i get the following error on the chrome console,
right away: Uncaught ReferenceError: require is not defined
The app.js file is referenced as a script indide of the index.html file
and the app.js file contains the definition for the require...
HELP!!! :(